Transfered props data in to outer files
