+++ app/src/components/Map/mapData.js
@@ -1,0 +1,7 @@
+const mapData = {
+  center: { lat: 49.97, lng: 36.25 },
+  zoom: 11,
+  bootstrapURLKeys: { key: 'AIzaSyB-7Fn2cGaiF3_2zTedyusFSiYHGn9RfMQ' }
+};
+
+export default mapData;
