+++ app/src/pages/Contacts/Contacts.js
@@ -3,36 +3,12 @@ import PropTypes from 'prop-types';
 /*import ContactsField from '../../components/ContactsField/ContactsField';
 import Map from '../../components/Map/Map';*/
 import './Contacts.scss';
+import dataContacts from './contactsFieldsData'
 import {
   Map,
   ContactsField
 } from 'Components';
 
-const dataContacts = [
-  {
-    id: 1,
-    Title: 'ADRESS',
-    data: '123, Rev Avenu, Hawai'
-  },
-
-  {
-    id: 2,
-    Title: 'PHONE',
-    data: '+81 2345854646'
-  },
-
-  {
-    id: 3,
-    Title: 'EMAIL',
-    data: 'email@email.com'
-  },
-
-  {
-    id: 4,
-    Title: 'STORE',
-    data: '123, Rev Avenu, Hawai'
-  }
-]
 
 class Contacts extends Component{
   constructor(props) {
