+++ app/src/pages/Contacts/contactsFieldsData.js
@@ -1,0 +1,28 @@
+const dataContacts = [
+  {
+    id: 1,
+    Title: 'ADRESS',
+    data: '123, Rev Avenu, Hawai'
+  },
+
+  {
+    id: 2,
+    Title: 'PHONE',
+    data: '+81 2345854646'
+  },
+
+  {
+    id: 3,
+    Title: 'EMAIL',
+    data: 'email@email.com'
+  },
+
+  {
+    id: 4,
+    Title: 'STORE',
+    data: '123, Rev Avenu, Hawai'
+  }
+]
+
+
+export default dataContacts
